refactor : data.dto.respose의 일부 하위폴더명 변경
