+++ data/src/main/java/com/knichu/data/dto/response/ShortWeatherResponseDTO.kt
@@ -1,7 +1,7 @@
 package com.knichu.data.dto.response
 
 import com.google.gson.annotations.SerializedName
-import com.knichu.data.dto.response.entity.short.SWResponse
+import com.knichu.data.dto.response.entity.shortWeather.SWResponse
 
 data class ShortWeatherResponseDTO(
     @field:SerializedName("response") val sWResponse: SWResponse? = null
