+++ data/src/main/java/com/knichu/data/dto/response/entity/liveWeather/LWItem.kt
@@ -1,4 +1,4 @@
-package com.knichu.data.dto.response.entity.live
+package com.knichu.data.dto.response.entity.liveWeather
 
 import com.google.gson.annotations.SerializedName
